requirements.txt
aiogram==3.14.0
aiosqlite==0.22.1
httpx==0.28.1
cryptography==46.0.5
pytz==2026.1.post1
python-dotenv==1.2.1
